requirements.txt
streamlit
base64
PIL
pandas
matplotlib.pyplot
sklearn.linear_model 
sklearn.model_selection 
sklearn.metrics 